added functions to reset game
Now there are functions like gtk_card_set_from_collection and game_field_clear that allow the game to be resetted once is finished.
A shuffle function was added as well to avoid having cards ordered by level.
A function that tells if the game is over or not has been added.